git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@49570 ebf7c1c2-ba36-0410-9fe8-c592906822b4
Former-commit-id: c0f0dce316146c70382b1cd63be8a5b47bb56f77

+++ src/jd/plugins/hoster/YoutubeDashV2.java
@@ -1702,7 +1702,7 @@ public class YoutubeDashV2 extends PluginForHost implements YoutubeHostPluginInt
                         if (previousChapter != null) {
                             final String title = (String) previousChapter.get("title");
                             final long startTimeStamp = ((Number) previousChapter.get("timeRangeStartMillis")).longValue();
-                            final int endTimeStamp = downloadLink.getIntegerProperty(YoutubeHelper.YT_DURATION, -1) * 1000;
+                            final long endTimeStamp = downloadLink.getIntegerProperty(YoutubeHelper.YT_DURATION, -1) * 1000l;
                             if (endTimeStamp > startTimeStamp) {
                                 final MetaDataEntry metaEntry = new MetaDataEntry("CHAPTER");
                                 metaEntry.put("TIMEBASE", "1/1000");
